+++ core/urls_hub.py
@@ -9,4 +9,6 @@ urlpatterns = [
     path("", views_hub.index, name="index"),
     path("ajouter/", views_hub.instance_add, name="instance_add"),
     path("<int:pk>/debrancher/", views_hub.instance_delete, name="instance_delete"),
+    path("<int:pk>/action/", views_hub.instance_action, name="instance_action"),
+    path("<int:pk>/cle-deblocage/", views_hub.instance_cle, name="instance_cle"),
 ]